Fix extreme lag while typing in 2500-line Elixir files.
Before: 1 char per second on an M1. After: no measurable lag.

Replace a non-greedy regex (which requires a lot of backtracking) with a greedy one.

Also tweak the `<indent>` subpattern to exclude newlines: this was causing some functions to be missed by the scanner, because the scanner would extend an earlier function until it happened to encounter a properly indented "end" with exactly the same pattern of newlines as it found above the function.